Add new Field Change the type of Images from several Images to one Image

+++ app/Http/Controllers/Admin/AdvertisementController.php
@@ -58,12 +58,13 @@ class AdvertisementController extends Controller
         $validator = Validator::make(
             $request->all(),
             [
-                'text_ar'               => 'required|string|min:30|max:255|unique:advertisement_translations,text',
-                'text_en'               => 'required|string|min:30|max:255|unique:advertisement_translations,text',
-                'user_id'             => 'required|string',
-                'section_id'            => 'required|string',
-                'category_id'           => 'required|string',
-                'image_name'            => 'required|image|mimes:jpg,jpeg,png,gif',
+                'text_ar'       => 'required|string|min:30|max:255|unique:advertisement_translations,text',
+                'text_en'       => 'required|string|min:30|max:255|unique:advertisement_translations,text',
+                'user_id'       => 'required|string',
+                'section_id'    => 'required|string',
+                'category_id'   => 'required|string',
+                'image_name'    => 'required|image|mimes:jpg,jpeg,png,gif',
+                'price'         => 'required',
             ]
         );
 
@@ -73,26 +74,20 @@ class AdvertisementController extends Controller
                 ->withInput();
         }
 
-        // Delete the Image From Public Folder
-        $Images = $advertisement->images()->pluck('image_name');
-        foreach ($Images as $Image) {
-            $destination = $Image;
+        if ($request->hasFile('image_name')) {
+            $destination = $advertisement->photo_name;
             if (File::exists($destination)) {
                 File::delete($destination);
             }
-        }
-
-        // Update The Image
-        if ($Images = $request->file('images')) {
-            foreach ($Images as $Image) {
-                $this->saveAdvertisementImage($Image, 'images/Advertisements', 600, 315, $advertisement);
-            }
+            $image_name = $this->saveImage($request->file('image_name'), 'images/Advertisements', 600, 315);
         }
 
         $data = [
-            'user_id'           => $request['user_id'],
-            'section_id'          => $request['section_id'],
-            'category_id'         => $request['category_id'],
+            'image_name'  => $image_name,
+            'price'       => $request['price'],
+            'user_id'     => $request['user_id'],
+            'section_id'  => $request['section_id'],
+            'category_id' => $request['category_id'],
             'ar' => [
                 'text'    => $request['text_ar'],
             ],
@@ -115,12 +110,9 @@ class AdvertisementController extends Controller
     public function destroy($id)
     {
         $advertisement = Advertisement::find($id);
-        $Images = $advertisement->images()->pluck('image_name');
-        foreach ($Images as $Image) {
-            $destination = $Image;
-            if (File::exists($destination)) {
-                File::delete($destination);
-            }
+        $destination = $advertisement->image_name;
+        if (File::exists($destination)) {
+            File::delete($destination);
         }
         $advertisement->delete();
         session()->flash('delete', 'تم حذف الإعلان بنجاح');
